Adding vllm and using it as a server

+++ src/routes/data.py
@@ -65,11 +65,36 @@ async def upload_data(request:Request, project_id:str, file:UploadFile,
     asset_resource = Asset(
         asset_project_id=project.id,
         asset_type=AssetTypeEnum.FILE.value,
-        asset_name=file_unique_id,
-        asset_size=os.path.getsize(file_path)
+        unique_asset_name=file_unique_id,
+        asset_size=os.path.getsize(file_path),
+        asset_name=file.filename
     )
     
     asset_record = await asset_model.create_asset(asset=asset_resource)
+    
+    old_asset_info, deleted_count = await asset_model.delete_asset_by_name(
+                                        asset_project_id=project.id,
+                                        asset_name=file.filename,
+                                        exclude_asset_id=asset_record.id  
+                                    )
+    
+    chunk_model = await ChunkModel.create_instance(
+                        db_client=request.app.db_client)
+    
+    # old_asset_ids = [id["id"] for id in old_asset_info]
+    
+    # old_asset_unique_name = [id["unique_asset_name"] for id in old_asset_info]
+    
+    for old_asset in old_asset_info:
+        await chunk_model.delete_chunk_by_asset_id(asset_id=old_asset["id"])
+        
+        old_file_path = os.path.join(project_dir_path, old_asset["unique_asset_name"])
+        try:
+            if os.path.exists(old_file_path):
+                os.remove(old_file_path)
+        except Exception as e:
+            logger.error(f"Error while deleting old file {old_file_path}: {e}")
+        
     
     return JSONResponse(content={"status":ResponseSignal.File_Upload_Success.value,
              "file_id":str(asset_record.id)
@@ -94,21 +119,23 @@ async def process_endpoint(request :Request, project_id:str, process_request:Pro
     asset_model = await AssetModel.create_instance(db_client=request.app.db_client)
     project_file_ids = {}
     
+    no_file_id = None
     if process_request.file_id:
         asset_record = await asset_model.get_asset_record(asset_project_id=project.id,
-                                                   asset_name=process_request.file_id)
+                                                   unique_asset_name=process_request.file_id)
         
         if asset_record is None:
             return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST, content={"signal":ResponseSignal.FILE_RECORD_ERROR.value})
             
         
-        project_file_ids = {asset_record.id:asset_record.asset_name}
+        project_file_ids = {asset_record.id:asset_record.unique_asset_name}
     
     else:
 
         project_files = await asset_model.get_all_project_assets(asset_project_id=project.id,
                                                                  asset_type=AssetTypeEnum.FILE.value)
-        project_file_ids = {record.id:record.asset_name for record in project_files}
+        project_file_ids = {record.id:record.unique_asset_name   for record in project_files}
+        no_file_id = True
     
     
     if len(project_file_ids)==0:
@@ -122,10 +149,18 @@ async def process_endpoint(request :Request, project_id:str, process_request:Pro
     chunk_model = await ChunkModel.create_instance(
     db_client=request.app.db_client)
     
-    if do_reset==1:
+    if do_reset==1 and no_file_id:
         _ = await chunk_model.delete_chunk_by_project_id(
             project_id=project.id
         )
+        
+    elif do_reset!=1 and no_file_id:
+        asset_ids = list(project_file_ids.keys())
+        unchunked_ids = [asset_id for asset_id in asset_ids if await  chunk_model.reversed_get_chunk_by_asset_id(asset_id=asset_id)]
+        un_chunked_files = [ project_file_ids[unchunked_id] for unchunked_id in unchunked_ids ]
+        project_file_ids =  dict(zip(unchunked_ids, un_chunked_files))
+        
+        
     for asset_id, file_id in project_file_ids.items():
         file_content = process_controller.get_file_content(file_id=file_id)
         
@@ -141,6 +176,8 @@ async def process_endpoint(request :Request, project_id:str, process_request:Pro
                                 content={
                                     "status":ResponseSignal.Processing_Failed.value
                                 })
+
+        await chunk_model.delete_chunk_by_asset_id(asset_id=asset_id)
 
         file_chunks_records = [
             DataChunk(
